pyproject.toml
[project]
name = "voice-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Maisha Tabassum", email = "maisha.tabassumm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "openai (>=1.88.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
